Restrict the size of input parameter.
git-svn-id: file:///research/zcentral/Vault/SVNROOT/ZFIN_WWW/trunk@5862 908b024a-c82b-635c-c293-f2b15b23277c

+++ lib/DB_functions/C/create_zero_pad_int.sql
@@ -1,4 +1,4 @@
-create function zero_pad_int(src int, width int)
+create function zero_pad_int(src int, width smallint)
   returning varchar(20)
   with (variant)
   external name
